rename package: navigations -> commands
